# internal/handlers/upload_check.go
package handlers

import (
	"strings"

	"github.com/gin-gonic/gin"

	"video-server/internal/response"
)

// @Summary Upload hash check
// @Tags upload
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param payload body UploadCheckRequest true "hash check payload"
// @Success 200 {object} APIResponse
// @Failure 200 {object} APIResponse
// @Router /upload/check [post]
func (a *API) UploadCheck(c *gin.Context) {
	var req struct {
		Hash     string `json:"hash"`
		FileSize int64  `json:"file_size"`
	}
	if err := c.ShouldBindJSON(&req); err != nil {
		bad(c, "invalid payload")
		return
	}
	req.Hash = strings.TrimSpace(req.Hash)
	if !isSHA256Hex(req.Hash) {
		bad(c, "invalid hash")
		return
	}
	if req.FileSize <= 0 {
		bad(c, "invalid file_size")
		return
	}

	videoID, exists, err := a.repo.FindVideoByHash(c.Request.Context(), req.Hash, req.FileSize)
	if err != nil {
		response.Error(c, 2, err.Error())
		return
	}
	if exists {
		ok(c, gin.H{"exists": true, "video_id": videoID})
		return
	}
	ok(c, gin.H{"exists": false})
}